pip-requirements.txt
django-cors-headers
djangorestframework-jsonapi
django-celery-results
django-polymorphic
